requirements.txt
python-dotenv
requests
tqdm
beautifulsoup4
playwright
duckduckgo-search
ddgs
pytube
youtube-transcript-api
selectolax
timezonefinder
timezonefinderL
geopy
quart
uvicorn
quart-cors
slowapi
yt-dlp
pytz
numba
loguru
sentence-transformers
faiss-cpu
aiohttp
torch
psutil